Initial commit 7b middleware example
