--- 7.1.middleware/deploy/ingress.yaml
@@ -1,14 +1,0 @@
-apiVersion: extensions/v1beta1
-kind: Ingress
-metadata:
-  annotations:
-    kubernetes.io/ingress.class: nginx
-  name: echo-ingress
-spec:
-  rules:
-    - http:
-        paths:
-          - backend:
-              serviceName: msgraph-dapr
-              servicePort: 80
-            path: /
