--- 7.1.middleware/deploy/msgraphpipeline.yaml
@@ -1,13 +1,0 @@
-apiVersion: dapr.io/v1alpha1
-kind: Configuration
-metadata:
-  name: msgraphpipeline
-spec:
-  tracing:
-    enabled: true
-    expandParams: true
-    includeBody: true
-  httpPipeline:
-    handlers:
-    - type: middleware.http.oauth2clientcredentials
-      name: msgraphsp
