--- 7.1.middleware/msgraphapp/Dockerfile
@@ -1,6 +1,0 @@
-FROM node:8-alpine
-WORKDIR /usr/src/app
-COPY . .
-RUN npm install
-EXPOSE 3000
-CMD [ "node", "app.js" ]
