+++ 7b.middleware-clientcredentials/msgraphapp/Dockerfile
@@ -1,0 +1,6 @@
+FROM node:8-alpine
+WORKDIR /usr/src/app
+COPY . .
+RUN npm install
+EXPOSE 3000
+CMD [ "node", "app.js" ]
