+++ 7b.middleware-clientcredentials/msgraphapp/package.json
@@ -1,0 +1,18 @@
+{
+  "name": "graphapp",
+  "version": "1.0.0",
+  "private": true,
+  "description": "",
+  "main": "app.js",
+  "scripts": {
+    "test": "echo \"Error: no test specified\" && exit 1"
+  },
+  "author": "",
+  "license": "ISC",
+  "dependencies": {
+    "body-parser": "^1.18.3",
+    "express": "^4.16.4",
+    "isomorphic-fetch": "^2.2.1",
+    "node-rest-client": "^3.1.0"
+  }
+}
